% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awsRekognition.R
\name{awsRekognition}
\alias{awsRekognition}
\title{Retrieve output from AWS Rekognition}
\usage{
awsRekognition(imagePath, AWS_ACCESS_KEY_ID = Sys.getenv("AWS_ACCESS_KEY_ID"),
  AWS_SECRET_ACCESS_KEY = Sys.getenv("AWS_SECRET_ACCESS_KEY"),
  AWS_BUCKET = Sys.getenv("AWS_BUCKET"),
  AWS_DEFAULT_REGION = Sys.getenv("AWS_DEFAULT_REGION"))
}
\arguments{
\item{imagePath}{path to image}

\item{AWS_ACCESS_KEY_ID}{defauts to Sys.getenv("AWS_ACCESS_KEY_ID")}

\item{AWS_SECRET_ACCESS_KEY}{defauts to Sys.getenv("AWS_SECRET_ACCESS_KEY")}

\item{AWS_BUCKET}{defauts to Sys.getenv("AWS_BUCKET")}

\item{AWS_DEFAULT_REGION}{defauts to Sys.getenv("AWS_DEFAULT_REGION")}
}
\description{
Retrieve output from AWS Rekognition
}
\examples{
imagePath <- system.file("ImageTests", "chimney_rock.jpg", package="WanderingEye")
awsRekognition(imagePath=imagePath,
               AWS_ACCESS_KEY_ID=Sys.getenv("AWS_ACCESS_KEY_ID"), 
               AWS_SECRET_ACCESS_KEY=Sys.getenv("AWS_SECRET_ACCESS_KEY"), 
               AWS_BUCKET=Sys.getenv("AWS_BUCKET"), 
               AWS_DEFAULT_REGION=Sys.getenv("AWS_DEFAULT_REGION"))
}
\keyword{image}
\keyword{processing}
