% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarifaiPredict.R
\name{clarifaiPredict}
\alias{clarifaiPredict}
\title{Retrieve output from Clarifai Image API}
\usage{
clarifaiPredict(imagePath, CLARIFAI_API_KEY = Sys.getenv("CLARIFAI_API_KEY"))
}
\arguments{
\item{imagePath}{local path, url path, google storage path}

\item{CLARIFAI_API_KEY}{defaults to Sys.getenv("CLARIFAI_API_KEY")}
}
\description{
Retrieve output from Clarifai Image API
}
\examples{
clarifaiPredict(imagePath="https://sports.cbsimg.net/images/blogs/nike-football.jpg")
imagePath <- system.file("ImageTests", "chimney_rock.jpg", package="WanderingEye")
clarifaiPredict(imagePath=imagePath)
}
\keyword{image}
\keyword{processing}
