% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleCloudVision.R
\name{googleCloudVision}
\alias{googleCloudVision}
\title{Retrieve output from Google Cloud Vision}
\usage{
googleCloudVision(imagePath, feature = "LABEL_DETECTION", numResults = 10,
  API_KEY = Sys.getenv("GCLOUD_VISION_API_KEY"))
}
\arguments{
\item{imagePath}{local path, url path, google storage path}

\item{feature}{Defaults to "LABEL_DETECTION". One of many features:
"LABEL_DETECTION","SAFE_SEARCH_DETECTION","LOGO_DETECTION",
"IMAGE_PROPERTIES","FACE_DETECTION","CROP_HINTS","DOCUMENT_TEXT_DETECTION",
"TEXT_DETECTION","WEB_DETECTION","LANDMARK_DETECTION"}

\item{numResults}{Value. Probably between 5 and 10. Defaults to 10.}

\item{API_KEY}{This defaults to Sys.getenv("GCLOUD_VISION_API_KEY")}
}
\description{
Retrieve output from Google Cloud Vision
}
\examples{
googleCloudVision(imagePath="https://sports.cbsimg.net/images/blogs/nike-football.jpg")
}
\keyword{image}
\keyword{processing}
