% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IBMWatsonVision.R
\name{IBMWatsonVision}
\alias{IBMWatsonVision}
\title{Retrieve output from IBM Watson Visual Recognition}
\usage{
IBMWatsonVision(imagePath, IBM_WATSON_API = Sys.getenv("IBM_WATSON_API"),
  IBM_WATSON_VERSION = Sys.getenv("IBM_WATSON_VERSION"))
}
\arguments{
\item{imagePath}{local path, url path, google storage path}

\item{IBM_WATSON_API}{This defaults to Sys.getenv("IBM_WATSON_API")}

\item{IBM_WATSON_VERSION}{This defaults to Sys.getenv("IBM_WATSON_VERSION")}
}
\description{
Retrieve output from IBM Watson Visual Recognition
}
\examples{
IBMWatsonVision(imagePath="https://sports.cbsimg.net/images/blogs/nike-football.jpg")
imagePath <- system.file("ImageTests", "chimney_rock.jpg", package="WanderingEye")
IBMWatsonVision(imagePath=imagePath)
}
\keyword{image}
\keyword{processing}
