% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awsProcess.R
\name{awsRekognition}
\alias{awsRekognition}
\title{Retrieve output from AWS Rekognition}
\usage{
awsRekognition(x)
}
\description{
Retrieve output from AWS Rekognition
}
\examples{
awsRekognition(x="chimney_rock.jpg")
}
\keyword{image}
\keyword{processing}
