% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/googleCloudVision.R
\name{googleCloudVision}
\alias{googleCloudVision}
\title{Retrieve output from Google Cloud Vision}
\usage{
googleCloudVision(imagePath, feature = "LABEL_DETECTION", numResults = 10,
  API_KEY = Sys.getenv("GCLOUD_VISION_API_KEY"))
}
\arguments{
\item{imagePath}{local path, url path, google storage path}

\item{feature}{Defaults to "LABEL_DETECTION". One of many features:
"LABEL_DETECTION","SAFE_SEARCH_DETECTION","LOGO_DETECTION",
"IMAGE_PROPERTIES","FACE_DETECTION","CROP_HINTS","DOCUMENT_TEXT_DETECTION",
"TEXT_DETECTION","WEB_DETECTION","LANDMARK_DETECTION"}

\item{numResults}{Value. Probably between 5 and 10. Defaults to 10.}

\item{API_KEY}{This defaults to Sys.getenv("GCLOUD_VISION_API_KEY")}
}
\description{
Retrieve output from Google Cloud Vision
}
\examples{
ImagePath1 <- "https://sports.cbsimg.net/images/blogs/nike-football.jpg"
ImagePath2 <- "http://ohscurrent.org/wp-content/uploads/2015/09/domus-01-google.jpg"
ImagePath3 <- "https://upload.wikimedia.org/wikipedia/commons/thumb/a/a1/Statue_of_Liberty_7.jpg/1200px-Statue_of_Liberty_7.jpg"
ImagePath4 <- system.file("ImageTests", "chimney_rock.jpg", package="WanderingEye")
ImagePath5 <- "http://a.espncdn.com/combiner/i?img=/i/headshots/nba/players/full/1966.png"
ImagePath6 <- system.file("ImageTests", "TextSample.png", package="WanderingEye")
googleCloudVision(imagePath=ImagePath1, feature="LABEL_DETECTION")
googleCloudVision(imagePath=ImagePath2, feature="LOGO_DETECTION")
googleCloudVision(imagePath=ImagePath3, feature="LANDMARK_DETECTION")
googleCloudVision(imagePath=ImagePath4, feature="WEB_DETECTION")
googleCloudVision(imagePath=ImagePath5, feature="FACE_DETECTION")
googleCloudVision(imagePath=ImagePath6, feature="TEXT_DETECTION")
googleCloudVision(imagePath=ImagePath6, feature="DOCUMENT_TEXT_DETECTION")
googleCloudVision(imagePath=ImagePath5, feature="SAFE_SEARCH_DETECTION")
googleCloudVision(imagePath=ImagePath5, feature="CROP_HINTS")
googleCloudVision(imagePath=ImagePath5, feature="IMAGE_PROPERTIES")
}
\keyword{image}
\keyword{processing}
