% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microsoftProcess.R
\name{microsoftComputerVision}
\alias{microsoftComputerVision}
\title{Retrieve output from Microsoft Computer Vision API}
\usage{
microsoftComputerVision(imagePath)
}
\arguments{
\item{imagePath}{local path, url path, google storage path}
}
\description{
Retrieve output from Microsoft Computer Vision API
}
\examples{
microsoftComputerVision(imagePath="https://sports.cbsimg.net/images/blogs/nike-football.jpg")
ImagePath <- system.file("ImageTests", "chimney_rock.jpg", package="WanderingEye")
microsoftComputerVision(imagePath=ImagePath)
}
\keyword{image}
\keyword{processing}
